Reworked all of the string-based pointer type conversion and checking to not involve casting to a long (problematic on 64-bit platforms). Instead, the functions work by scanning the raw bytes of the pointer value itself and packing the value to a string.
Caveat: This causes the pointer "values" as seen in the target
language to be reversed on little-endian architectures (e.g., the
mathematical ordering of the hexadecimal digits is wrong).  Since
pointer values in SWIG are *supposed* to be opaque, this is probably
not a big deal.  However, it will break any extensions that might have
tried to manipulate the pointer strings manually.  It could also
make debugging interesting if you were expecting the SWIG pointer
value to exactly match addresses in the debugger.

Bonus: The packing/unpacking functions can work with any binary data
up to about 500 bytes.  Therefore, it might be possible to pack small
objects that are larger than a single pointer (e.g., C++ pointers to
member functions).

Unknown:  Are pointers to functions the same size as void * on all
important platforms?    The direct packing/unpacking may bypass magical
function pointer casts as performed by the C compiler (if any).


git-svn-id: svn://svn.code.sf.net/p/swig/code/branches/rel-1-3@1521 d7a574c2-d0a9-4346-9435-0907474fee8b