Added support for data-member --> member function transformation. This is actually a long requested feature that mostly relates to this problem:
Suppose you have a structure like this:

   struct Foo {
      int bar;
      int spam;
   };

Now, suppose that you really wanted to access the attributes
as functions instead of the normal SWIG behavior (possibly
to conform to a naming convention or to make something
appear more class-like). For example:

    f = Foo()
    a = f->getBar()        # instead of f.bar
    f->setSpam(42)         # instead of f.spam = 42

In the past, this was extremely problematic--you had to go write a
bunch of helper functions or hack something with %addmethods.
Now it's easy.  Just do this:

   %attributefunc(get%(title)s,set%(title)s)
   struct Foo {
      int bar;
      int spam;
   };
   %noattributefunc

The arguments to %attributefunc are simply printf-style format strings
where %s is the data member name.  These strings can also use the
special DOH encoding methods (upper,lower,title) as shown above.

Note: this directive is similar to the new %callback directive.


git-svn-id: svn://svn.code.sf.net/p/swig/code/branches/rel-1-3@1620 d7a574c2-d0a9-4346-9435-0907474fee8b