Fixed SF [ #458466 ] Bug 1.3.7 class scoped enumerations.
When default arguments are used, SWIG now performs a symbol table lookup to
see if anything about the value is known.  If so, a fully qualified scope-name is
used.  For example, in this code:

class Foo {
   enum speed { FAST, SLOW };
   void blah(speed s = FAST);
};

the default value of 's' gets set to Foo::FAST instead of FAST.

This resolution is currently performed in the parser.  I'm not sure if this is the
best place to do it, but I couldn't think of a better alternative.


git-svn-id: svn://svn.code.sf.net/p/swig/code/branches/rel-1-3@1692 d7a574c2-d0a9-4346-9435-0907474fee8b