To better deal with operators and funny symbols, a new language API method validIdentifier() has been added.  SWIG will not wrap a declaration unless this function returns a 1 for the symbol table name (the name used in the scripting interface).
A language module can use this to tell the core that unrenamed
operators are okay by simply checking identifiers to see if they start
with 'operator ' and returning a 1 if so. Note: other parts of the
module would need to be prepared for dealing with the actual operators.

Also: removed the is_multiple_definition() method.

Don't forget to do a 'make clean', 'make' after this change.


git-svn-id: svn://svn.code.sf.net/p/swig/code/branches/rel-1-3@1743 d7a574c2-d0a9-4346-9435-0907474fee8b