Fixed very subtle semantic problem in type system.  If you have a type like this:
    Foo * const

The function SwigType_ispointer() would return 0.  Yet, this type is very clearly a
pointer (albeit a qualified one).   This has the potential to cause problems with typemaps
and certain language module features.   The ispointer() function now returns 1--even if
the pointer is qualified.

Other changes:

SwigType_del_pointer() will delete the pointer and any qualifiers attached to it.
SwigType_isfunction() returns 1 even if a function is qualified.  This change is
needed to better support 'const' methods in classes.  For example:

class Foo {
   ...
   void blah() const;
   ...
};

Note: const and non-const methods can be overloaded separately so we're going to need this.

Added SwigType_pop_function() to return a function declarator and any associated qualifiers.


git-svn-id: svn://svn.code.sf.net/p/swig/code/branches/rel-1-3@1767 d7a574c2-d0a9-4346-9435-0907474fee8b