Added backend symbol table handling.   This has long been a hole in SWIG. For example:
   Foo *new_Foo();
   class Foo {
   };

In this case, no error message was generated even though the creation of a
constructor for Foo generates a name clash with new_Foo().   Two new functions
in the Language class, addSymbol() and lookupSymbol() can be used to provide a
diagnostic message.   I've modified the Python, Tcl, Perl, Guile, Mzscheme, and PHP
modules with the necessary code.  I'm not sure about Ruby and Java.

Note: This change is needed to make the wrapping of namespaces more sane--since
they are flattened in the target language, name clashes are not detected in the normal
way.


git-svn-id: svn://svn.code.sf.net/p/swig/code/branches/rel-1-3@2563 d7a574c2-d0a9-4346-9435-0907474fee8b