Change to %feature directive.
Background: A major use of %feature is to set flags.  For example:

    %feature("new","1")  foo;
    %feature("immutable","1") foo;
    %feature("classic","1") foo;
    ...

However, in order to set a flag, you also have to specify the extra "1"--which is
easy to forget (or at least odd).   To fix this, %feature has been changed to
implicitly set a value of "1" if none is supplied.  For example:

    %feature("new") foo;           // Implicitly sets feature:new = "1"
    %feature("immutable") foo;
    %feature("class") foo;

Since this conflicts with the way in which features are cleared, the following
declaration is now used to clear a feature:

    %feature("new","") foo;       // Clear feature:new

(We might define a macro %clearfeature to do the same thing).

Comments:  It's more common to set features than to clear them---I think we
should make this syntax easy.


git-svn-id: svn://svn.code.sf.net/p/swig/code/branches/rel-1-3@3451 d7a574c2-d0a9-4346-9435-0907474fee8b