Fix directors + protected pure virtual members + protected constructor.
Now it doesn't crash, you don't need to use dirprot, and no, it is not
emitting all the protected members, unless dirprot is used.

Swig only emits the protected members or constructors needed to prevent
the compilation for crashing. The rest of the protected members
are emitted when dirprot is used.


git-svn-id: svn://svn.code.sf.net/p/swig/code/trunk/SWIG@6493 d7a574c2-d0a9-4346-9435-0907474fee8b