Use -O2 in the GCC examples as GCC defaults to not optimising at all.
Document using GCC's -fPIC option as the standard approach - x86 is the oddity
here and most architectures require it (or are always PIC), and on x86 it's
better to use it or else code pages from the library need relocations and can't
be shared.

Use "python2.5" rather than "python2.0" in paths in example commands.

Fix a typo.


git-svn-id: svn://svn.code.sf.net/p/swig/code/trunk@11334 d7a574c2-d0a9-4346-9435-0907474fee8b