Correct explanation of how to match on class name in %rename.
Replace incorrect documentation of $parentNode from %rename discussion: it
advised using match$parentNode but this doesn't work because the parent node
is not yet set when %rename is parsed.

Document the "fullname" attribute of %rename which can be used to restrict the
match to the given full name of a declaration only.

git-svn-id: svn://svn.code.sf.net/p/swig/code/trunk@12291 d7a574c2-d0a9-4346-9435-0907474fee8b