D compiler sanity checks in configure; check D2 by default if no D1 compiler can be found.
For both D1 and D2, if the compiler has not been specified explicitly via a command line option, the configure script tries to compile a minimal program stub. Only if this succeeds, the respective compiler is enabled.

If neither a working D1 nor a working D2 compiler has been found, D is skipped completely as usual.

git-svn-id: svn://svn.code.sf.net/p/swig/code/trunk@12334 d7a574c2-d0a9-4346-9435-0907474fee8b