Fix bug which could result in %rename not taking effect for derived classes.
We used to modify the hash table that we iterated on in
Swig_name_object_inherit() and this could, and sometimes did, change the
iteration order in such way that not all entries we were looking for could be
found. In practice this means that sometimes the methods renamed or ignored in
the base class could be mysteriously not renamed or ignored in a derived
class.

Fix this by avoiding modifying the hash table in place and using another
temporary hash table instead.

git-svn-id: svn://svn.code.sf.net/p/swig/code/trunk@12865 d7a574c2-d0a9-4346-9435-0907474fee8b